♻️ GH-243 Enforce uniform Result[T] contract at MCP boundary
The SuccessResult.to_dict() polymorphic isinstance branch silently
changed output shape, record_upgrade bypassed the Result contract,
_gh_api leaked CompletedProcess to 20+ call sites, and server_cli.py
had grown to 1834 lines with ~60 handlers. ADR-0009 records the
uniform-shape decision, which is breaking-change-adjacent.

Solution:
- A4+B5: drop the to_dict isinstance branch; the value is a Mapping
  returned unchanged; add @runtime_checkable ResultProtocol
- H1+I9: extract record_upgrade() -> Result[dict] into the domain
  module; the MCP handler is now a thin .to_dict() wrapper
- I8: _gh_api returns Result[dict] via _parse_gh_api_result, with
  _gh_api_raw serving graphql/jq raw CompletedProcess callers
- A6: split server_cli.py into github/git/plan/audit/misc tool
  modules; enumerate_mcp and the cwd guard now scan all of them
- Record ADR-0009 and cover the new contract with tests

Fixes: GH-243
