+++ src/dev10x/domain/common/result.py
@@ -1,7 +1,21 @@
 from __future__ import annotations
 
+from collections.abc import Mapping
 from dataclasses import dataclass, field
-from typing import Any
+from typing import Any, Protocol, cast, runtime_checkable
+
+
+@runtime_checkable
+class ResultProtocol(Protocol):
+    """Contract every value crossing the MCP boundary must satisfy.
+
+    Both ``SuccessResult`` and ``ErrorResult`` implement ``to_dict``.
+    The ``@server.tool()`` boundary asserts ``isinstance(x,
+    ResultProtocol)`` to catch a handler that forgot ``.to_dict()``
+    before the wire-encode step (ADR-0009).
+    """
+
+    def to_dict(self) -> dict[str, Any]: ...
 
 
 @dataclass(frozen=True)
@@ -9,9 +23,10 @@ class SuccessResult[T]:
     value: T
 
     def to_dict(self) -> dict[str, Any]:
-        if isinstance(self.value, dict):
-            return self.value
-        return {"value": self.value}
+        # Contract (ADR-0009): a SuccessResult that reaches the MCP
+        # boundary wraps a Mapping, returned unchanged. Internal-only
+        # results (e.g. Result[RepositoryRef]) never call to_dict().
+        return dict(cast("Mapping[str, Any]", self.value))
 
 
 @dataclass(frozen=True)
