+++ tests/github/test_github.py
@@ -49,7 +49,7 @@ class TestPrCommentsResolveSingle:
         )
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_resolves_single_comment(
         self,
         mock_api: AsyncMock,
@@ -71,7 +71,7 @@ class TestPrCommentsResolveSingle:
         assert mock_api.call_count == 2
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_converts_int_comment_id_to_string(
         self,
         mock_api: AsyncMock,
@@ -131,7 +131,7 @@ class TestPrCommentsResolveBatch:
         )
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_resolves_multiple_comments_in_two_calls(
         self,
         mock_api: AsyncMock,
@@ -157,7 +157,7 @@ class TestPrCommentsResolveBatch:
         assert "r2" in result.value["data"]
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_batch_query_uses_aliased_nodes(
         self,
         mock_api: AsyncMock,
@@ -180,7 +180,7 @@ class TestPrCommentsResolveBatch:
         assert "n2:" in query_str
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_comment_ids_takes_precedence_over_comment_id(
         self,
         mock_api: AsyncMock,
@@ -210,7 +210,7 @@ class TestPrCommentsResolveBatch:
 
 class TestPrCommentsResolveErrors:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_when_query_fails(
         self,
         mock_api: AsyncMock,
@@ -230,7 +230,7 @@ class TestPrCommentsResolveErrors:
         assert result.error == "GraphQL error"
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_when_thread_not_found(
         self,
         mock_api: AsyncMock,
@@ -249,7 +249,7 @@ class TestPrCommentsResolveErrors:
         assert "Could not find thread" in result.error
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_when_thread_id_invalid(
         self,
         mock_api: AsyncMock,
@@ -270,7 +270,7 @@ class TestPrCommentsResolveErrors:
         assert "Could not find thread" in result.error
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_partial_failure_includes_warnings(
         self,
         mock_api: AsyncMock,
@@ -304,7 +304,7 @@ class TestPrCommentsResolveErrors:
         assert any("PRRC_bad" in w for w in result.value["warnings"])
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_mutation_error_returns_error(
         self,
         mock_api: AsyncMock,
@@ -328,7 +328,7 @@ class TestPrCommentsResolveErrors:
 
 class TestResolveReviewThreadDirect:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_resolves_by_thread_ids(
         self,
         mock_api: AsyncMock,
@@ -351,7 +351,7 @@ class TestResolveReviewThreadDirect:
         assert "must start with PRRT_" in result.error
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_resolves_by_comment_ids(
         self,
         mock_api: AsyncMock,
@@ -383,7 +383,7 @@ class TestResolveReviewThreadDirect:
 
 class TestPrCommentListFilters:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_filters_by_review_id(
         self,
         mock_api: AsyncMock,
@@ -406,10 +406,12 @@ class TestPrCommentListFilters:
         )
 
         assert isinstance(result, SuccessResult)
-        assert {c["id"] for c in result.value} == {1, 3}
+        # ADR-0009: list payloads are wrapped under "value" to satisfy the
+        # Mapping contract; the {"value": [...]} wire shape is unchanged.
+        assert {c["id"] for c in result.value["value"]} == {1, 3}
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_all_when_no_review_id(
         self,
         mock_api: AsyncMock,
@@ -427,10 +429,27 @@ class TestPrCommentListFilters:
         result = await gh.pr_comments(action="list", pr_number=42)
 
         assert isinstance(result, SuccessResult)
-        assert len(result.value) == 2
+        assert len(result.value["value"]) == 2
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
+    async def test_non_json_output_passes_through_as_dict(
+        self,
+        mock_api: AsyncMock,
+        mock_resolve_repo: AsyncMock,
+    ) -> None:
+        # When gh returns non-JSON, _parse_gh_api_result yields a
+        # {"raw_output": ...} dict; _pr_comment_list passes that dict
+        # through unwrapped (ADR-0009 Mapping contract already met).
+        mock_api.return_value = _completed(stdout="not json")
+
+        result = await gh.pr_comments(action="list", pr_number=42)
+
+        assert isinstance(result, SuccessResult)
+        assert result.value == {"raw_output": "not json"}
+
+    @pytest.mark.asyncio
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_unresolved_only_uses_graphql(
         self,
         mock_api: AsyncMock,
@@ -524,7 +543,7 @@ class TestMinimizeComments:
         )
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_batches_into_single_request(
         self,
         mock_api: AsyncMock,
@@ -545,7 +564,7 @@ class TestMinimizeComments:
         assert "classifier: OUTDATED" in query
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_accepts_alternate_classifier(
         self,
         mock_api: AsyncMock,
@@ -581,7 +600,7 @@ class TestMinimizeComments:
         assert "Invalid classifier" in result.error
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_on_api_failure(
         self,
         mock_api: AsyncMock,
@@ -608,7 +627,7 @@ class TestResolveRepo:
 
 class TestPrCommentReply:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_posts_reply(
         self,
         mock_api: AsyncMock,
@@ -634,7 +653,7 @@ class TestPrCommentReply:
         assert call_kwargs["repo"] == "owner/repo"
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_on_api_failure(
         self,
         mock_api: AsyncMock,
@@ -654,7 +673,7 @@ class TestPrCommentReply:
         assert isinstance(result, ErrorResult)
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_coerces_numeric_string_comment_id_to_int(
         self,
         mock_api: AsyncMock,
@@ -672,7 +691,7 @@ class TestPrCommentReply:
         assert mock_api.call_args.kwargs["fields"]["in_reply_to"] == 3130499018
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_rejects_non_numeric_comment_id(
         self,
         mock_api: AsyncMock,
@@ -691,7 +710,7 @@ class TestPrCommentReply:
 
 class TestPrIssueComment:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_posts_top_level_comment(
         self,
         mock_api: AsyncMock,
@@ -717,7 +736,7 @@ class TestPrIssueComment:
         assert call_kwargs["repo"] == "owner/repo"
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_on_api_failure(
         self,
         mock_api: AsyncMock,
@@ -738,7 +757,7 @@ class TestPrIssueComment:
 
 class TestPrCommentsActionReply:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_coerces_numeric_string_to_int(
         self,
         mock_api: AsyncMock,
@@ -776,7 +795,7 @@ class TestPrCommentsActionReply:
 
 class TestPrCommentsActionEdit:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_patches_comment_body(
         self,
         mock_api: AsyncMock,
@@ -799,7 +818,7 @@ class TestPrCommentsActionEdit:
         assert mock_api.call_args.kwargs["as_bot"] is True
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_coerces_numeric_string_to_int(
         self,
         mock_api: AsyncMock,
@@ -843,7 +862,7 @@ class TestPrCommentsActionEdit:
 
 class TestRequestReview:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_requests_user_reviewers(
         self,
         mock_api: AsyncMock,
@@ -863,7 +882,7 @@ class TestRequestReview:
         assert fields["reviewers"] == ["alice"]
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_requests_team_reviewers(
         self,
         mock_api: AsyncMock,
@@ -897,7 +916,7 @@ class TestPrCommentsStrategyDispatch:
         assert "get, list, reply, edit, resolve" in result.error
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_get_action_requires_comment_id(
         self,
         mock_api: AsyncMock,
@@ -909,7 +928,7 @@ class TestPrCommentsStrategyDispatch:
         assert "comment_id required" in result.error
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_list_action_requires_pr_number(
         self,
         mock_api: AsyncMock,
@@ -921,7 +940,7 @@ class TestPrCommentsStrategyDispatch:
         assert "pr_number required" in result.error
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_get_action_fetches_comment(
         self,
         mock_api: AsyncMock,
@@ -936,7 +955,7 @@ class TestPrCommentsStrategyDispatch:
         assert isinstance(result, SuccessResult)
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_list_action_fetches_comments(
         self,
         mock_api: AsyncMock,
@@ -951,7 +970,7 @@ class TestPrCommentsStrategyDispatch:
         assert isinstance(result, SuccessResult)
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_reply_action_posts_comment(
         self,
         mock_api: AsyncMock,
@@ -995,7 +1014,7 @@ class TestPrCommentsStrategyDispatch:
 
 class TestMilestoneClose:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_closes_milestone(
         self,
         mock_api: AsyncMock,
@@ -1017,7 +1036,7 @@ class TestMilestoneClose:
         assert call.kwargs["fields"] == {"state": "closed"}
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_when_repo_unresolved(
         self,
         mock_api: AsyncMock,
@@ -1029,7 +1048,7 @@ class TestMilestoneClose:
         mock_api.assert_not_awaited()
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_on_api_failure(
         self,
         mock_api: AsyncMock,
@@ -1045,7 +1064,7 @@ class TestMilestoneClose:
 
 class TestMilestoneCreate:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_creates_milestone(
         self,
         mock_api: AsyncMock,
@@ -1070,7 +1089,7 @@ class TestMilestoneCreate:
         assert call.kwargs["fields"]["description"] == "desc"
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_on_api_failure(
         self,
         mock_api: AsyncMock,
@@ -1351,7 +1370,7 @@ class TestIssueList:
 
 class TestUpdatePr:
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_updates_body(
         self,
         mock_api: AsyncMock,
@@ -1373,7 +1392,7 @@ class TestUpdatePr:
         assert call.kwargs["fields"] == {"body": "new body"}
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_updates_title_and_base(
         self,
         mock_api: AsyncMock,
@@ -1401,7 +1420,7 @@ class TestUpdatePr:
         assert "at least one" in result.error.lower()
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_when_repo_unresolved(
         self,
         mock_api: AsyncMock,
@@ -1413,7 +1432,7 @@ class TestUpdatePr:
         mock_api.assert_not_awaited()
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_returns_error_on_api_failure(
         self,
         mock_api: AsyncMock,
@@ -1427,7 +1446,7 @@ class TestUpdatePr:
         assert "422" in result.error
 
     @pytest.mark.asyncio
-    @patch("dev10x.github._gh_api", new_callable=AsyncMock)
+    @patch("dev10x.github._gh_api_raw", new_callable=AsyncMock)
     async def test_uses_explicit_repo_when_provided(
         self,
         mock_api: AsyncMock,
